cmake: support shareware and full version
The Wolfenstein 3D game has a shareware version and a full version
build. Each of these require asset files that are shipped with the
game. This commit adds definitions that can be used to explicitly
define the version to build and where to find and copy asset files.

To build the full version, define asset path with
CPP_WOLF_FULL_ASSETS_DIR. For shareware define
CPP_WOLF_SHAREWARE_ASSETS_DIR.